requirements.txt
discord.py
anthropic
pydantic
python-dotenv
pytz
python-frontmatter
rapidfuzz
